added formula to convert from AS8510 ADC readings to milliamps

+++ Apps/Src/Amps.c
@@ -48,7 +48,7 @@ void Amperes_Post(void) {
 }
 #endif
 
-static int16_t latestMeasureMilliAmps;
+static int32_t latestMeasureMilliAmps;
 
 /** Amps_Init
  * Initializes hardware to begin current monitoring.
